Convert cipher files to enigma files
